add macros.rs
