+++ src/connector/src/base.rs
@@ -31,7 +31,7 @@ use crate::nexmark::source::reader::NexmarkSplitReader;
 use crate::nexmark::{NexmarkSplit, NexmarkSplitEnumerator};
 use crate::pulsar::source::reader::PulsarSplitReader;
 use crate::pulsar::{PulsarEnumeratorOffset, PulsarSplit, PulsarSplitEnumerator};
-use crate::ConnectorProperties;
+use crate::{ConnectorProperties, impl_split, impl_split_enumerator, impl_split_reader};
 
 /// [`SplitEnumerator`] fetches the split metadata from the external source service.
 /// NOTE: It runs in the meta server, so probably it should be moved to the `meta` crate.
@@ -90,6 +90,31 @@ pub enum SplitEnumeratorImpl {
     Nexmark(NexmarkSplitEnumerator),
 }
 
+impl_split_enumerator! {
+    [ ] ,
+    { Kafka, KafkaSplitEnumerator },
+    { Pulsar, PulsarSplitEnumerator },
+    { Kinesis, KinesisSplitEnumerator },
+    { Nexmark, NexmarkSplitEnumerator }
+}
+
+impl_split! {
+    [ ] ,
+    { Kafka, KAFKA_SPLIT_TYPE, KafkaSplit },
+    { Pulsar, PULSAR_SPLIT_TYPE, PulsarSplit },
+    { Kinesis, KINESIS_SPLIT_TYPE, KinesisSplit },
+    { Nexmark, NEXMARK_SPLIT_TYPE, NexmarkSplit }
+}
+
+impl_split_reader! {
+    [ ] ,
+    { Kafka, KafkaSplitReader },
+    { Pulsar, PulsarSplitReader },
+    { Kinesis, KinesisMultiSplitReader },
+    { Nexmark, NexmarkSplitReader },
+    { Dummy, DummySplitReader }
+}
+
 pub type DataType = risingwave_common::types::DataType;
 
 #[derive(Clone, Debug)]
@@ -114,113 +139,7 @@ pub trait SplitMetaData: Sized {
     fn restore_from_bytes(bytes: &[u8]) -> Result<Self>;
 }
 
-macro_rules! impl_split_enumerator {
-    ([], $({ $variant_name:ident, $split_enumerator_name:ident} ),*) => {
-        impl SplitEnumeratorImpl {
 
-             pub async fn create(properties: ConnectorProperties) -> Result<Self> {
-                match properties {
-                    $( ConnectorProperties::$variant_name(props) => $split_enumerator_name::new(props).await.map(Self::$variant_name), )*
-                    other => Err(anyhow!("split enumerator type for config {:?} is not supported", other)),
-                }
-             }
-
-             pub async fn list_splits(&mut self) -> Result<Vec<SplitImpl>> {
-                match self {
-                    $( Self::$variant_name(inner) => inner.list_splits().await.map(|ss| ss.into_iter().map(SplitImpl::$variant_name).collect_vec()), )*
-                }
-             }
-        }
-    }
-}
-
-impl_split_enumerator! {
-            [ ] ,
-            { Kafka, KafkaSplitEnumerator },
-            { Pulsar, PulsarSplitEnumerator },
-            { Kinesis, KinesisSplitEnumerator },
-            { Nexmark, NexmarkSplitEnumerator }
-}
-
-macro_rules! impl_split {
-    ([], $({ $variant_name:ident, $split_name:ident, $split:ty} ),*) => {
-        impl SplitImpl {
-            pub fn id(&self) -> String {
-                match self {
-                    $( Self::$variant_name(inner) => inner.id(), )*
-                }
-            }
-
-            pub fn to_json_bytes(&self) -> Bytes {
-                match self {
-                    $( Self::$variant_name(inner) => inner.encode_to_bytes(), )*
-                }
-            }
-
-            pub fn get_type(&self) -> String {
-                match self {
-                    $( Self::$variant_name(_) => $split_name, )*
-                }
-                    .to_string()
-            }
-
-            pub fn restore_from_bytes(split_type: String, bytes: &[u8]) -> Result<Self> {
-                match split_type.as_str() {
-                    $( $split_name => <$split>::restore_from_bytes(bytes).map(Self::$variant_name), )*
-                    other => Err(anyhow!("split type {} not supported", other)),
-                }
-            }
-        }
-    }
-}
-
-impl_split! {
-            [ ] ,
-            { Kafka, KAFKA_SPLIT_TYPE, KafkaSplit },
-            { Pulsar, PULSAR_SPLIT_TYPE, PulsarSplit },
-            { Kinesis, KINESIS_SPLIT_TYPE, KinesisSplit },
-            { Nexmark, NEXMARK_SPLIT_TYPE, NexmarkSplit }
-}
-
-macro_rules! impl_split_reader {
-    ([], $({ $variant_name:ident, $split_reader_name:ident} ),*) => {
-        impl SplitReaderImpl {
-            pub async fn next(&mut self) -> Result<Option<Vec<SourceMessage>>> {
-                match self {
-                    $( Self::$variant_name(inner) => inner.next().await, )*
-                }
-            }
-
-             pub async fn create(
-                config: ConnectorProperties,
-                state: ConnectorStateV2,
-                columns: Option<Vec<Column>>,
-            ) -> Result<Self> {
-                if let ConnectorStateV2::Splits(s) = &state {
-                    if s.is_empty() {
-                        return Ok(Self::Dummy(Box::new(DummySplitReader {})));
-                    }
-                }
-
-                let connector = match config {
-                     $( ConnectorProperties::$variant_name(props) => Self::$variant_name(Box::new($split_reader_name::new(props, state, columns).await?)), )*
-                    _ => todo!()
-                };
-
-                Ok(connector)
-            }
-        }
-    }
-}
-
-impl_split_reader! {
-            [ ] ,
-            { Kafka, KafkaSplitReader },
-            { Pulsar, PulsarSplitReader },
-            { Kinesis, KinesisMultiSplitReader },
-            { Nexmark, NexmarkSplitReader },
-            { Dummy, DummySplitReader }
-}
 
 /// The persistent state of the connector.
 #[derive(Debug, Clone, Serialize, Deserialize)]
